[backend] Reduced network load by only transmitting diffs on save.
The current network load was too high for saving a whole file every
second.  Since a it is unrealistic for a person to edit their entire
file in that time, we compute a unified diff of the changes they've
made and then send only that over the network.  On the server side, we
fork to start the `patch' program which adjusts the file so it is
correct.

We've copied the script from jsdifflib <http://snowtide.com/jsdifflib>
into the VCS to enable us to compute the diffs on the client machine.
We've also added functionality for it to output properly formatted
diffs.  The only new outside dependency is that the `patch' program
must be installed on the host machine.